+++ CHANGELOG.md
@@ -24,7 +24,15 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [2.0.0] - Dev branch
+## [2.0.1] - Dev branch
+
+### Added
+
+- [service]: Add path of service file.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
+## [2.0.0] - 2026-07-01
 
 ### Added